fix: Arrumando os IDs no microserviço de transação

+++ transacao-service/src/main/java/br/com/solutis/transacao_service/service/TransacaoService.java
@@ -1,6 +1,6 @@
 package br.com.solutis.transacao_service.service;
 
-import br.com.solutis.transacao_service.dto.transacao.TransacaoResumedResponseDto;
+import br.com.solutis.transacao_service.dto.TransacaoResumedResponseDto;
 import br.com.solutis.transacao_service.mapper.TransacaoMapper;
 import br.com.solutis.transacao_service.repository.TransacaoRepository;
 import lombok.AllArgsConstructor;
@@ -18,7 +18,7 @@ public class TransacaoService {
     @Autowired
     private TransacaoRepository repository;
 
-    public List<TransacaoResumedResponseDto> listarPorIdCartao(Integer id) {
+    public List<TransacaoResumedResponseDto> listarPorIdCartao(Long id) {
         if (id == null) return null;
 
         List<TransacaoResumedResponseDto> transacoes = repository.findAllByCartaoId(id).stream().map(mapper::toResumedResponse).toList();
